requirements.txt
Django==4.2.13
python-dotenv==1.0.1
mysqlclient==2.2.1
daphne==4.1.2
uuid==1.30
django-debug-toolbar==4.4.6
django-constance==4.0.0
slack-sdk==3.34.0
channels==4.1.0